llm/requirements.txt
supabase>=2.4.0
python-dotenv>=1.0.1
httpx>=0.27.0
pydantic>=2.8.0
rich>=13.7.1
